pyproject.toml
[tool.poetry]
name = "Etna"
version = "0.1.0"
description = ""
authors = ["Dan Bentley <dan@numiko.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gunicorn = "^20.1.0"
elasticsearch = "^7.12.0"
wagtail = "^2.12.3"
psycopg2 = "^2.8.6"
platformshconfig = "^2.4.0"
wagtailmedia = "^0.7.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
ipdb = "^0.13.7"
django-extensions = "^3.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
